# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "File": {
        "properties": {
          "access": {
            "type": "string"
          },
          "content_type": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "digest": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "path": {
            "type": "string"
          },
          "size": {
            "type": [
              "integer",
              "null"
            ]
          },
          "state": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "version_count": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "path",
          "state",
          "access",
          "content_type",
          "metadata",
          "version_count",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "FilePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/File"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "Opaque keyset cursor for the next page; null when drained.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "type": "object"
      },
...
      "Run": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "ended_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "output": {
            "type": [
              "object",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "workflow": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "workflow",
          "status",
          "created_at"
        ],
        "type": "object"
      },
      "RunPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Run"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "Opaque keyset cursor for the next page; null when drained.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "header": {
        "in": "header",
        "name": "x-copal-tenant",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "copal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/files": {
      "get": {
        "operationId": "list_files",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous page's next_cursor.",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by state (indexed).",
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sort order; every value is backed by an index.",
            "in": "query",
            "name": "sort",
            "schema": {
              "enum": [
                "created_at",
                "-created_at"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilePage"
                }
              }
            },
            "description": "Page of resources."
          }
        },
        "x-requires-scopes": [
          "read"
        ]
      },
      "post": {
        "description": "Create the file record; bytes follow through the content upload or an upload grant.",
        "operationId": "create_files",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "access": {
                    "enum": [
                      "public",
                      "private",
                      "tenant",
                      "grant"
                    ],
                    "type": "string"
                  },
                  "content_type": {
                    "type": "string"
                  },
                  "idempotency_key": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  },
                  "path": {
                    "type": "string"
                  }
                },
                "required": [
                  "path"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            },
            "description": "The resource."
          }
        },
        "x-requires-scopes": [
          "write"
        ]
      }
    },
...
    "/v1/runs": {
      "get": {
        "operationId": "list_runs",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous page's next_cursor.",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by status (indexed).",
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sort order; every value is backed by an index.",
            "in": "query",
            "name": "sort",
            "schema": {
              "enum": [
                "created_at",
                "-created_at"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            },
            "description": "Page of resources."
          }
        },
        "x-requires-scopes": [
          "read"
        ]
      },
      "post": {
        "description": "Start a workflow run.",
        "operationId": "start_runs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string"
                  },
                  "idempotency_key": {
                    "type": "string"
                  },
                  "input": {
                    "type": "object"
                  },
                  "mode": {
                    "type": "string"
                  },
                  "workflow": {
                    "type": "string"
                  }
                },
                "required": [
                  "workflow"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Action result."
          }
        },
        "x-requires-scopes": [
          "write"
        ]
      }
    },
    "/v1/runs/{id}": {
      "get": {
        "operationId": "get_runs",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "The resource."
          },
          "404": {
            "description": "Not found."
          }
        },
        "x-requires-scopes": [
          "read"
        ]
      }
    },
...
  },
  "security": [
    {
      "header": []
    }
  ],
...
}